requirements.txt
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cffi==1.13.2
coverage==4.5.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==3.0
django-classy-tags==0.9.0
django-crispy-forms==1.8.1
django-enumfield==2.0
django-extensions==2.2.5
django-guardian==2.1.0
django-tailwind==0.6.0
entrypoints==0.3
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
zipp==0.6.0
